requirements.txt
streamlit
pandas
numpy
matplotlib
numpy-financial
openpyxl
scipy
google-generativeai
python-dotenv